+++ source/rrSelectionRecord.cpp
@@ -249,7 +249,7 @@ using namespace std;
 
 
 
-SelectionRecord::SelectionRecord(const int& _index,
+SelectionRecord::SelectionRecord(const size_t& _index,
         const SelectionType _type, const string& _p1,
         const string& _p2) :
         index(_index), p1(_p1), p2(_p2), selectionType(_type)
